+++ src/css/global/_mixins.scss
@@ -78,3 +78,19 @@
     font-weight: bold;
     width: max-content;
 }
+
+@mixin scroll-bar {
+    &::-webkit-scrollbar {
+        width: 0.5rem;
+        height: 0.5rem;
+    }
+
+    &::-webkit-scrollbar-track {
+        background: var(--scrollbar-track-bg-color);
+    }
+
+    &::-webkit-scrollbar-thumb {
+        background-color: var(--scrollbar-thumb-bg-color);
+        border-radius: 0.375rem;
+    }
+}
